When the tool is running don't check the tool status more than once every 2 secs.
When initially checking on the tool status, if we get a network
error, mark thge network as down.

When we check the card status, if we get a reply from the
network, mark the network as working.

Only report toolusage if the network is running.

N.B. compiled and tested with energia 15.