Try to fix vairous network timeout and packet mangling issues.
possibly more visible now with netverbose off, since the with it on the serial.print() things add serendepitious delays which hid the issues.

The symptom was that POSTing to /tooluse/ would return -1 about one time in 20 or so, usually beacause only the "200 OK" part was read from the network and the rest discarded.