add doorbell functionality, and software timed release functionality.
Software timed release: Connect PM_0 pin to GND via a Normally Open
push button switch.
This will open the door for the usual timeout feature, and send an EXIT
notification.

Doorbell function:
Standard pushbutton as per acnode to PM_7 pin and GND.
Will send a BELL message and blink LED in an orangey way.

Responding to BELL is left as an exercise for the reader.