Make more robust in cases of network link failure
When the network link goes down, the doorbot tries
to do stuff anyway, which results in a generally nasty situation
where it will probably hit enough timeouts to trigger the watchdog.
This is bad because hitting the watchdog timer clears the cache, meaning
people who otherwise would have been permitted access now are not.

This commit adds checking of link status to a few key places. It's only
a start, but it should help in the short term while a more thorough examination
is performed.