Stop the EEProm cache from "borrowing" storage from the config
The eeprom cache iteration logic contained a while(1) and no regard for the end of the eeprom. This results
in it wrapping around, and trying to use the config as a card record, which, with validation, ends up modifying the
config to make it no longer config. This is most unhelpful.

This commit adds actual bounds validation to the EEProm cache logic. Novel idea, I know.