Add option to invert colours
Current hardware requires driving the LED lines low to turn them on. The
code therefore specifies all colours inverted.

This method of driving the active low pins makes little sense, and more
to the point, is incompatible with the newer hardware design, which uses
buffer transistors to allow an on-board LED to be driven in parallel
without loading the external LED.

This commit fixes this insanity by making the inversion configurable
(default on for compatibility) and defining the colours normally.